pyproject.toml
[project]
name = "riverwood"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.2",
    "chromadb>=1.3.4",
    "elevenlabs>=2.22.0",
    "fastapi>=0.121.0",
    "google-generativeai>=0.8.5",
    "groq>=0.33.0",
    "indic-transliteration>=2.3.75",
    "openai>=2.7.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
]
